requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pinecone-client==2.2.4
huggingface-hub==0.22.2
streamlit==1.31.1
scikit-learn==1.4.0
python-dotenv==1.0.1
numpy==1.26.4
spacy-transformers
selenium>=4.15.2
pandas>=2.1.3
webdriver-manager>=4.0.1
urllib3>=2.0.7
google-generativeai>=0.3.1